Updated latest SE binaries. Fixed unselecting item in main window. Fixed issue in calculating ore in super large asteroids. Split out my experimental code from test code. Rewrote the display of ship component breakdown, and fixed some of the calculations which were putting out wrong values (seriously undervalued).

+++ Main/SEToolbox/SEToolbox/ViewModels/GenerateVoxelFieldViewModel.cs
@@ -307,7 +307,7 @@ namespace SEToolbox.ViewModels
                     asteroid.Load(voxelDesign.VoxelFile.SourceFilename, voxelDesign.MainMaterial.Value);
 
                     IList<byte> baseAssets;
-                    Dictionary<byte, int> materialVoxelCells;
+                    Dictionary<byte, long> materialVoxelCells;
 
                     asteroid.CalculateMaterialCellAssets(out baseAssets, out materialVoxelCells);
